+++ app/Http/Controllers/ExtraController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Extra;
 use Illuminate\Http\Request;
 use App\Http\Requests\ExtraRequest;
+use Illuminate\Support\Facades\Auth;
 
 class ExtraController extends Controller
 {
@@ -41,6 +42,7 @@ class ExtraController extends Controller
             'price'=> $request->price,
             'photo' => $request->photo ? $request->file('photo')->store('public/photos') : null,
             'description' => $request->description,
+            'user_id' => Auth::user()->id,
         ]);
         return redirect(route('extra.index'))->with('extraCreated', 'Hai inserito correttamente un accessorio');
     }
@@ -57,7 +59,10 @@ class ExtraController extends Controller
      * Show the form for editing the specified resource.
      */
     public function edit(Extra $extra)
-    {
+    {   
+        if($extra->user_id != Auth::id()){
+            return redirect(route('extra.index'))->with('accessDenied', 'non sei autorizzato ad effetture modifiche a questo post');
+        }
         return view ('extra.edit', compact('extra'));
     }
 
@@ -90,6 +95,10 @@ class ExtraController extends Controller
      */
     public function destroy(Extra $extra)
     {
+        if($extra->user_id != Auth::id()){
+            return redirect(route('bike.index'))->with('accessDenied', 'Non sei autorizzato ad effetturare questa operazione');
+        }
+
         $extra->delete();
         return redirect(route('extra.index'))->with('extraDelete', 'Hai cancellato correttamente il tuo annuncio');
     }
